fix: build.sh

+++ build.sh
@@ -29,6 +29,8 @@ echo_content() {
 }
 
 init_var() {
+  trojan_panel_version=latest
+
   arch_arr=('amd64' 'arm64')
 
   touch Dockerfile
@@ -36,7 +38,10 @@ init_var() {
 
 main() {
   for get_arch in ${arch_arr[*]}; do
-    echo_content blue "开始构建Trojan Panel ${get_arch}"
+    if [[ ! -f build/trojan-panel-linux-${get_arch} ]]; then
+      continue
+    fi
+    echo_content blue "开始构建trojan-panel-linux-${get_arch}"
 
     cat >Dockerfile <<EOF
 FROM alpine:3.15
@@ -65,11 +70,31 @@ ENTRYPOINT ./trojan-panel \
     -redisPassword=\${redis_pass}
 EOF
 
-    docker buildx build --platform linux/"${get_arch}" -t jonssonyan/trojan-panel-"${get_arch}" .
+    docker buildx build --platform linux/"${get_arch}" -t jonssonyan/trojan-panel-linux-"${get_arch}" .
     if [[ "$?" == "0" ]]; then
-      echo_content green "Trojan Panel ${get_arch}构建成功"
+      echo_content green "trojan-panel-linux-${get_arch}构建成功"
+      echo_content blue "开始推送trojan-panel-linux-${get_arch}"
+      docker image tag jonssonyan/trojan-panel-linux-"${get_arch}":latest jonssonyan/trojan-panel:latest && \
+      docker image push jonssonyan/trojan-panel:latest && \
+      docker rmi -f jonssonyan/trojan-panel:latest
+      if [[ "$?" == "0" ]]; then
+        echo_content green "镜像名称：jonssonyan/trojan-panel:latest 架构：${get_arch}推送成功"
+      else
+        echo_content red "镜像名称：jonssonyan/trojan-panel:latest 架构：${get_arch}推送失败"
+      fi
+
+      if [[ ${trojan_panel_version != "latest"} ]]; then
+        docker image tag jonssonyan/trojan-panel-linux-"${get_arch}":latest jonssonyan/trojan-panel:${trojan_panel_version} && \
+        docker image push jonssonyan/trojan-panel:${trojan_panel_version} && \
+        docker rmi -f jonssonyan/trojan-panel:${trojan_panel_version}
+        if [[ "$?" == "0" ]]; then
+          echo_content green "镜像名称：jonssonyan/trojan-panel:${trojan_panel_version} 架构：${get_arch}推送成功"
+        else
+          echo_content green "镜像名称：jonssonyan/trojan-panel:${trojan_panel_version} 架构：${get_arch}推送成功"
+        fi
+      fi
     else
-      echo_content red "Trojan Panel ${get_arch}构建失败"
+      echo_content red "trojan-panel-linux-${get_arch}构建失败"
     fi
   done
 }
